Use shared tags component for listings
