+++ app/javascript/shared/components/tags.jsx
@@ -19,6 +19,7 @@ class Tags extends Component {
     this.state = {
       selectedIndex: -1,
       searchResults: [],
+      additionalTags: [],
       cursorIdx: 0,
       prevLen: 0,
     };
@@ -37,6 +38,19 @@ class Tags extends Component {
       .split(',')
       .map(item => item !== undefined && item.trim())
       .filter(item => item.length > 0);
+  }
+
+  componentDidMount() {
+    if (this.props.listingTags === true) {
+      this.setState({
+        additionalTags: {jobs: ['remote', 'remoteoptional', 'lgbtbenefits', 'greencard', 'senior', 'junior', 'intermediate', '401k', 'fulltime', 'contract', 'temp'],
+                        forhire: ['remote', 'remoteoptional', 'lgbtbenefits', 'greencard', 'senior', 'junior', 'intermediate', '401k', 'fulltime', 'contract', 'temp'],
+                        forsale: ['laptop', 'desktopcomputer', 'new', 'used'],
+                        events: ['conference', 'meetup'],
+                        collabs: ['paid', 'temp']
+                        }
+      })
+    }
   }
 
   componentDidUpdate() {
@@ -79,7 +93,8 @@ class Tags extends Component {
           type="text"
           ref={t => (this.textArea = t)}
           className={`${this.props.classPrefix}__tags`}
-          placeholder="tags"
+          placeholder={`${this.props.maxTags} tags max, comma separated, no spaces or special characters`}
+          autoComplete={this.props.autoComplete || 'on'}
           value={this.props.defaultValue}
           onInput={this.handleInput}
           onKeyDown={this.handleKeyDown}
@@ -173,6 +188,17 @@ class Tags extends Component {
         filters: 'supported:true',
       })
       .then(content => {
+        if (this.props.listingTags === true) {
+          const { additionalTags } = this.state
+          const { category } = this.props
+          const additionalItems = (additionalTags[category] || []).filter(t => (t.indexOf(query) > -1))
+          const resultsArray = content.hits;
+          additionalItems.forEach(t => {
+            if (resultsArray.indexOf(t) === -1) {
+              resultsArray.push({name: t});
+            }
+          })
+        }
         this.setState({
           searchResults: content.hits.filter(
             hit => !this.selected.includes(hit.name),
